requirements.txt
python-binance
pandas
ta
numpy